+++ python/lsst/ts/m2/__init__.py
@@ -34,4 +34,5 @@ from .csc import *
 from .mock_message_telemetry import *
 from .mock_message_event import *
 from .mock_command import *
+from .mock_model import *
 from .mock_server import *